sell-master/src/main/java/com/cheng/sell/service/impl/CategoryServiceImpl.java
package com.cheng.sell.service.impl;

import com.cheng.sell.dataobject.ProductCategory;
import com.cheng.sell.repository.ProductCategoryRepository;
import com.cheng.sell.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created with IntelliJ IDEA.
 * Description:
 * 类目服务实现
 * @author cheng
 * Date: 2018-07-01
 * Time: 下午6:59
 */
@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    public CategoryServiceImpl() {
    }

    @Override
    public ProductCategory findById(Integer categoryId) {
        return repository.findById(categoryId).get();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return repository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
